backend/requirements.txt
fastapi
uvicorn
pandas
llama-cpp-python
